Rename troubleshoot publication file
