+++ publication_troubleshoot.py
